Split stone into seperate files for classes
